Refactor hint_font function location
This commit relocates the hint_font function from the afdko_tools.py to the autohint.py module. This change also includes importing necessary modules for the hint_font function to the autohint.py module and removing it from afdko_tools.py list of exports.
