+++ foundrytools_cli_2/lib/otf/afdko_tools.py
@@ -1,13 +1,10 @@
 import typing as t
 from contextlib import contextmanager
-from pathlib import Path
 
-from afdko.otfautohint.__main__ import ACOptions, _validate_path
-from afdko.otfautohint.autohint import FontInstance, fontWrapper, openFont
 from cffsubr import desubroutinize, subroutinize
 from fontTools.ttLib import TTFont
 
-__all__ = ["cff_subr", "cff_desubr", "hint_font"]
+__all__ = ["cff_subr", "cff_desubr"]
 
 
 @contextmanager
@@ -49,26 +46,3 @@ def cff_desubr(font: TTFont) -> None:
     _validate_sfnt_version(font=font)
     with _restore_flavor(font):
         desubroutinize(otf=font)
-
-
-def hint_font(
-    in_file: Path,
-    options: ACOptions,
-) -> TTFont:
-    """
-    Applies hinting to an OpenType-PS font file and returns the hinted TTFont object.
-
-    Parameters:
-        in_file: Path to the font file.
-        options: An ACOptions object containing the options to use for hinting.
-
-    Returns:
-        TTFont: A hinted TTFont object.
-    """
-
-    in_file = _validate_path(in_file)
-    font = openFont(in_file, options=options)
-    font_instance = FontInstance(font=font, inpath=in_file, outpath=None)
-    fw = fontWrapper(options=options, fil=[font_instance])
-    fw.hint()
-    return fw.fontInstances[0].font.ttFont
